+++ code/app/src/main/res/navigation/nav_graph.xml
@@ -3,26 +3,25 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/nav_graph"
-    app:startDestination="@id/FirstFragment">
+    app:startDestination="@id/BalanceFragment">
 
     <fragment
-        android:id="@+id/FirstFragment"
-        android:name="com.example.arcmobile.FirstFragment"
-        android:label="@string/first_fragment_label"
-        tools:layout="@layout/fragment_first">
+        android:id="@+id/BalanceFragment"
+        android:name="com.example.arcmobile.BalanceFragment"
+        android:label="balance_fragment_label"
+        tools:layout="@layout/fragment_reload">
 
         <action
-            android:id="@+id/action_FirstFragment_to_SecondFragment"
-            app:destination="@id/SecondFragment" />
+            android:id="@+id/action_BalanceFragment_to_ReloadFragment"
+            app:destination="@id/ReloadFragment" />
     </fragment>
     <fragment
-        android:id="@+id/SecondFragment"
-        android:name="com.example.arcmobile.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second">
+        android:id="@+id/ReloadFragment"
+        android:name="com.example.arcmobile.ReloadFragment"
+        android:label="reload_fragment_label">
 
         <action
-            android:id="@+id/action_SecondFragment_to_FirstFragment"
-            app:destination="@id/FirstFragment" />
+            android:id="@+id/action_ReloadFragment_to_BalanceFragment"
+            app:destination="@id/BalanceFragment" />
     </fragment>
 </navigation>
